requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
streamlit
requests
